pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "taan-grpo"
version = "0.1.0"
description = "TAAN-enhanced GRPO multi-task RL training system with Type-Aware Advantage Normalization"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "datasets>=2.18",
    "pyyaml>=6.0",
    "accelerate>=0.30",
    "wandb>=0.17",
]

[project.optional-dependencies]
train = [
    "vllm>=0.6.0",
    "flash-attn>=2.5",
    "deepspeed>=0.14",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
taan-train = "scripts.train:main"
taan-eval  = "scripts.eval:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
